+++ xwiki-platform-core/xwiki-platform-attachment/xwiki-platform-attachment-picker/xwiki-platform-attachment-picker-macro/src/main/resources/META-INF/components.txt
@@ -1,0 +1,1 @@
+org.xwiki.attachment.picker.internal.AttachmentPickerMacro
