+++ xwiki-platform-core/xwiki-platform-attachment/xwiki-platform-attachment-picker/xwiki-platform-attachment-picker-war/pom.xml
@@ -1,0 +1,81 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<!--
+ * See the NOTICE file distributed with this work for additional
+ * information regarding copyright ownership.
+ *
+ * This is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation; either version 2.1 of
+ * the License, or (at your option) any later version.
+ *
+ * This software is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this software; if not, write to the Free
+ * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
+ * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
+-->
+
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+  <parent>
+    <groupId>org.xwiki.platform</groupId>
+    <artifactId>xwiki-platform-attachment-picker</artifactId>
+    <version>14.4-SNAPSHOT</version>
+  </parent>
+  <artifactId>xwiki-platform-attachment-picker-war</artifactId>
+  <name>XWiki Platform - Attachment - Picker - WAR</name>
+  <packaging>war</packaging>
+  <!-- We need this module because we're missing a skin extension plugin to load JavaScript code from a WebJar.
+    See XWIKI-12788 (Introduce skin extension plugins for webjar resources) -->
+  <description>Velocity macros and JavaScript files that can't be packaged in the Attachment Picker WebJar.</description>
+  <build>
+    <plugins>
+      <!-- Javascript compression -->
+      <plugin>
+        <groupId>com.github.blutorange</groupId>
+        <artifactId>closure-compiler-maven-plugin</artifactId>
+        <executions>
+          <execution>
+            <goals>
+              <goal>minify</goal>
+            </goals>
+          </execution>
+        </executions>
+        <configuration>
+          <!-- Fix the path to the source file from the generated source map. This works ATM because all the JavaScript
+            source files are located in the same folder. Note that the source mapping is not going to be perfect if the
+            JavaScript code contains Velocity code. -->
+          <closureSourceMapLocationMappings>
+            <closureSourceMapLocationMapping>
+              <name>uicomponents/widgets/</name>
+              <!-- We can't leave the value empty because it generates "null" in the path. -->
+              <value>./</value>
+            </closureSourceMapLocationMapping>
+          </closureSourceMapLocationMappings>
+        </configuration>
+      </plugin>
+      <plugin>
+        <groupId>org.apache.maven.plugins</groupId>
+        <artifactId>maven-war-plugin</artifactId>
+        <configuration>
+          <failOnMissingWebXml>false</failOnMissingWebXml>
+          <webResources>
+            <!-- Include License resources -->
+            <resource>
+              <directory>${project.build.directory}/maven-shared-archive-resources/META-INF</directory>
+              <targetPath>META-INF</targetPath>
+            </resource>
+          </webResources>
+          <!-- Exclude the LICENSE/NOTICE files from the classes/ directory since we're including them in META-INF/ at
+               the root of the WAR -->
+          <packagingExcludes>WEB-INF/**</packagingExcludes>
+        </configuration>
+      </plugin>
+    </plugins>
+  </build>
+</project>
