--- xwiki-platform-core/xwiki-platform-image/xwiki-platform-image-picker/xwiki-platform-image-picker-macro/src/main/resources/META-INF/components.txt
@@ -1,1 +1,0 @@
-org.xwiki.image.picker.internal.ImagePickerMacro
