--- xwiki-platform-core/xwiki-platform-image/xwiki-platform-image-picker/xwiki-platform-image-picker-war/src/main/webapp/resources/uicomponents/widgets/imagePicker.css
@@ -1,33 +1,0 @@
-.imageGroup {
-  display: inline-block;
-  max-width: 150px;
-  text-overflow: ellipsis;
-  overflow: hidden;
-  white-space: nowrap;
-  margin: 5px;
-}
-
-.imageGroup a {
-  color: inherit;
-}
-
-.imageGroup img {
-  object-fit: contain;
-  max-width: 150px;
-  max-height: 150px;
-  margin-bottom: 5px;
-}
-
-.imageGroup > div {
-  position: relative;
-}
-
-.imageGroup input[type='radio'] {
-  position: absolute;
-  top: 0;
-  left: 0;
-}
-
-.imagePickerSearch {
-  margin-bottom: 1em;
-}
