+++ xwiki-platform-core/xwiki-platform-web/xwiki-platform-web-war/pom.xml
@@ -59,7 +59,7 @@
     </dependency>
     <dependency>
       <groupId>org.xwiki.platform</groupId>
-      <artifactId>xwiki-platform-image-picker-war</artifactId>
+      <artifactId>xwiki-platform-attachment-picker-war</artifactId>
       <version>${project.version}</version>
       <type>war</type>
     </dependency>
